Improved performance with Kotlin Coroutines, added map features, cleaned code, added utility methods and classes. Added intelligent tracking methods. Added location management for user. Added new bus data. Added setting option for intelligent tracking (default ON). Updated UI.

+++ app/src/main/java/com/example/intelligentbustracker/service/LocationService.kt
@@ -335,10 +335,13 @@ class LocationService : Service() {
             removeUpdatesTask.addOnSuccessListener {
                 running = false
                 Common.setRequestingLocationUpdates(this, false)
-                currentUserDocumentReference?.delete()
-                currentUserDocumentReference = null
-                currentUser = null
-                stopSelf()
+                currentUserDocumentReference?.delete()?.addOnSuccessListener {
+                    currentUserDocumentReference = null
+                    currentUser = null
+                    stopSelf()
+                }?.addOnFailureListener {
+                    Toast.makeText(this@LocationService, "Failed to remove user reference from Firestore.", Toast.LENGTH_LONG).show()
+                }
             }.addOnFailureListener {
                 Toast.makeText(this@LocationService, "Failed to remove location updates.", Toast.LENGTH_LONG).show()
             }
@@ -389,13 +392,16 @@ class LocationService : Service() {
     override fun onTaskRemoved(rootIntent: Intent?) {
         // Delete current user data
         if (!Common.requestingLocationUpdates(this)) {
-            val removeUpdatesTask = fusedLocationProviderClient!!.removeLocationUpdates(locationCallback!!)
-            removeUpdatesTask.addOnSuccessListener {
-                currentUserDocumentReference?.delete()
-                stopSelf()
-                super.onTaskRemoved(rootIntent)
-            }.addOnFailureListener {
-                Toast.makeText(this@LocationService, "Failed to remove location updates.", Toast.LENGTH_LONG).show()
+            if (running) {
+                val removeUpdatesTask = fusedLocationProviderClient!!.removeLocationUpdates(locationCallback!!)
+                removeUpdatesTask.addOnSuccessListener {
+                    currentUserDocumentReference?.delete()
+                    running = false
+                    stopSelf()
+                    super.onTaskRemoved(rootIntent)
+                }.addOnFailureListener {
+                    Toast.makeText(this@LocationService, "Failed to remove location updates.", Toast.LENGTH_LONG).show()
+                }
             }
         }
     }
